pyproject.toml
[project]
name = "slack-bridge"
version = "0.1.0"
description = "AWS Lambda with Slack integration for agent workflows"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.175.0",
    "constructs>=10.0.0,<11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
